backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2
django-redis==5.4.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
whitenoise==6.7.0
python-decouple==3.8
redis==5.0.4
Pillow==10.3.0
